fixed the bug and camera works now
