+++ app/src/main/java/com/example/needlevision/MainActivity.java
@@ -43,10 +43,7 @@ import java.util.List;
 
 public class MainActivity extends AppCompatActivity {
     private static final int RC_SIGN_IN = 1;
-    private static final int CAMERA_PERM_CODE = 101;
-    private static final int CAMERA_REQUEST_CODE = 102;
-    private static final int PHOTO_REQUEST_CODE = 103;
-    private String photoPath;
+
     GoogleSignInClient mGoogleSignInClient;
     SignInButton signInButton;
 
@@ -79,11 +76,10 @@ public class MainActivity extends AppCompatActivity {
             loadLoginPage();
             getSupportActionBar().hide();
         } else {
-            // Signed in successfully, show authenticated UI.
-            Intent intent = new Intent(MainActivity.this, PostLoginActivity.class);
-            startActivity(intent);
-
-//            loadPagerPage();
+            // Signed in successfully, show authenticated UI. DELETE LATER
+//            Intent intent = new Intent(MainActivity.this, PostLoginActivity.class);
+//            startActivity(intent);
+            loadPager();
         }
     }
     
@@ -93,7 +89,7 @@ public class MainActivity extends AppCompatActivity {
             // Load the loadPagerPage function
             @Override
             public void onClick(View v) {
-                loadPagerPage();
+                loadPager();
             }
         });
         findViewById(R.id.sign_in_button).setOnClickListener(new View.OnClickListener() {
@@ -104,82 +100,6 @@ public class MainActivity extends AppCompatActivity {
                 signIn();
             }
         });
-    }
-
-    // Loads the Fragment Slides
-    private void loadPagerPage(){
-        setContentView(R.layout.activity_main);
-        List<Fragment> list = new ArrayList<>();
-        list.add(new map_fragment());
-        list.add(new posts_fragment());
-
-        pager = findViewById(R.id.pager);
-        pagerAdapter = new SlidePagerAdapter(getSupportFragmentManager(), list);
-        pager.setAdapter(pagerAdapter);
-
-        loadFAB();
-    }
-
-    private void loadFAB(){
-        FloatingActionButton fab = findViewById(R.id.fab);
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Snackbar.make(view, "Bring Up Camera", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
-                checkCameraPermissions();
-            }
-        });
-    }
-
-    private void checkCameraPermissions() {
-        // Display camera permission prompt
-        if (ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
-            ActivityCompat.requestPermissions(this, new String[] { Manifest.permission.CAMERA }, CAMERA_PERM_CODE);
-        } else {
-            //Open Camera
-            openCamera();
-        }
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        if (requestCode == CAMERA_PERM_CODE) {
-            if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                openCamera();
-            } else {
-                Toast.makeText(this, "Camera Permission is Required to Use camera.", Toast.LENGTH_SHORT).show();
-            }
-        }
-    }
-
-    private void openCamera() {
-        Intent camera = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-        if (camera.resolveActivity(this.getPackageManager()) != null) {
-            File photoFile = null;
-            photoFile = createImageFile();
-            if (photoFile != null) {
-                Uri photoURI = FileProvider.getUriForFile(this, "com.example.android.FileProvider", photoFile);
-                camera.putExtra(MediaStore.EXTRA_OUTPUT, photoURI);
-                startActivityForResult(camera, CAMERA_REQUEST_CODE);
-            }
-        }
-    }
-
-    // Create Image File
-    public File createImageFile() {
-        String timestamp = new SimpleDateFormat("yyyy-MM-dd").format(new Date());
-        String imageName = timestamp + "_";
-        File storageDirectory = this.getExternalFilesDir(Environment.DIRECTORY_PICTURES);
-        File image = null;
-
-        try {
-            image = File.createTempFile(imageName, ".jpg", storageDirectory);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        photoPath = image.getAbsolutePath();
-        return image;
     }
 
     private void signIn() {
@@ -197,18 +117,6 @@ public class MainActivity extends AppCompatActivity {
             // a listener.
             Task<GoogleSignInAccount> task = GoogleSignIn.getSignedInAccountFromIntent(data);
             handleSignInResult(task);
-        } else if (requestCode == CAMERA_REQUEST_CODE) {
-            if (resultCode != RESULT_OK) {
-                File nullPhoto = new File(photoPath);
-                nullPhoto.delete();
-            } else {
-                Intent photoActivity = new Intent(MainActivity.this, PhotoActivity.class);
-                photoActivity.putExtra("PHOTO_PATH", photoPath);
-                startActivityForResult(photoActivity, PHOTO_REQUEST_CODE);
-            }
-        } else if (requestCode == PHOTO_REQUEST_CODE) {
-            //Return from the photo activity
-            //Refresh map pins/list view entries?
         }
     }
 
@@ -216,9 +124,10 @@ public class MainActivity extends AppCompatActivity {
         try {
             GoogleSignInAccount account = completedTask.getResult(ApiException.class);
 
+            loadPager();
             // Signed in successfully, show authenticated UI.
-            Intent intent = new Intent(MainActivity.this, PostLoginActivity.class);
-            startActivity(intent);
+//            Intent intent = new Intent(MainActivity.this, PostLoginActivity.class);
+//            startActivity(intent);
         } catch (ApiException e) {
             // The ApiException status code indicates the detailed failure reason.
             // Please refer to the GoogleSignInStatusCodes class reference for more information.
@@ -226,11 +135,16 @@ public class MainActivity extends AppCompatActivity {
         }
     }
 
-    private void socialMediaUpload() {
-        Intent mediaUpload = new Intent();
-        mediaUpload.setAction(Intent.ACTION_SEND);
-        mediaUpload.putExtra(Intent.EXTRA_STREAM, FileProvider.getUriForFile(this, "com.example.android.FileProvider", new File(photoPath)));
-        mediaUpload.setType("image/jpg");
-        this.startActivity(Intent.createChooser(mediaUpload, this.getResources().getText(R.string.send_to)));
+    private void loadPager(){
+        Intent intent = new Intent(MainActivity.this, PostActivity.class);
+        startActivity(intent);
     }
+
+//    private void socialMediaUpload() {
+//        Intent mediaUpload = new Intent();
+//        mediaUpload.setAction(Intent.ACTION_SEND);
+//        mediaUpload.putExtra(Intent.EXTRA_STREAM, FileProvider.getUriForFile(this, "com.example.android.FileProvider", new File(photoPath)));
+//        mediaUpload.setType("image/jpg");
+//        this.startActivity(Intent.createChooser(mediaUpload, this.getResources().getText(R.string.send_to)));
+//    }
 }
