+++ app/src/main/java/com/example/needlevision/PostActivity.java
@@ -1,16 +1,23 @@
 package com.example.needlevision;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.core.content.FileProvider;
 import androidx.fragment.app.Fragment;
 import androidx.viewpager.widget.PagerAdapter;
 import androidx.viewpager.widget.ViewPager;
 
-import android.annotation.SuppressLint;
+import android.Manifest;
 import android.content.Intent;
-import android.content.res.TypedArray;
+import android.content.pm.PackageManager;
+import android.net.Uri;
 import android.os.Bundle;
+import android.os.Environment;
+import android.provider.MediaStore;
 import android.view.View;
-import android.widget.ImageButton;
+import android.widget.Toast;
 
 import com.example.needlevision.adapters.SlidePagerAdapter;
 import com.example.needlevision.fragments.map_fragment;
@@ -18,10 +25,19 @@ import com.example.needlevision.fragments.posts_fragment;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.snackbar.Snackbar;
 
+import java.io.File;
+import java.io.IOException;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 public class PostActivity extends AppCompatActivity {
+    private static final int CAMERA_PERM_CODE = 101;
+    private static final int CAMERA_REQUEST_CODE = 102;
+    private static final int PHOTO_REQUEST_CODE = 103;
+    private String photoPath;
+
     private ViewPager pager;
     private PagerAdapter pagerAdapter;
 
@@ -51,12 +67,83 @@ public class PostActivity extends AppCompatActivity {
             public void onClick(View view) {
                 Snackbar.make(view, "Bring Up Camera", Snackbar.LENGTH_LONG)
                         .setAction("Action", null).show();
+                checkCameraPermissions();
+
             }
         });
     }
 
     public void setActionBarTitle(String title) {
         getSupportActionBar().setTitle(title);
+    }
+
+    public void checkCameraPermissions() {
+        // Display camera permission prompt
+        if (ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
+            ActivityCompat.requestPermissions(this, new String[] { Manifest.permission.CAMERA }, CAMERA_PERM_CODE);
+        } else {
+            //Open Camera
+            openCamera();
+        }
+    }
+
+    public void openCamera() {
+        Intent camera = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
+        if (camera.resolveActivity(this.getPackageManager()) != null) {
+            File photoFile = null;
+            photoFile = createImageFile();
+            if (photoFile != null) {
+                Uri photoURI = FileProvider.getUriForFile(this, "com.example.android.FileProvider", photoFile);
+                camera.putExtra(MediaStore.EXTRA_OUTPUT, photoURI);
+                startActivityForResult(camera, CAMERA_REQUEST_CODE);
+            }
+        }
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+
+        if (requestCode == CAMERA_REQUEST_CODE) {
+            if (resultCode != RESULT_OK) {
+                File nullPhoto = new File(photoPath);
+                nullPhoto.delete();
+            } else {
+                Intent photoActivity = new Intent(this, PhotoActivity.class);
+                photoActivity.putExtra("PHOTO_PATH", photoPath);
+                startActivityForResult(photoActivity, PHOTO_REQUEST_CODE);
+            }
+        } else if (requestCode == PHOTO_REQUEST_CODE) {
+            //Return from the photo activity
+            //Refresh map pins/list view entries?
+        }
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        if (requestCode == CAMERA_PERM_CODE) {
+            if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                openCamera();
+            } else {
+                Toast.makeText(this, "Camera Permission is Required to Use camera.", Toast.LENGTH_SHORT).show();
+            }
+        }
+    }
+
+    // Create Image File
+    public File createImageFile() {
+        String timestamp = new SimpleDateFormat("yyyy-MM-dd").format(new Date());
+        String imageName = timestamp + "_";
+        File storageDirectory = this.getExternalFilesDir(Environment.DIRECTORY_PICTURES);
+        File image = null;
+
+        try {
+            image = File.createTempFile(imageName, ".jpg", storageDirectory);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        photoPath = image.getAbsolutePath();
+        return image;
     }
